ATLP-51 Impala support

+++ src/main/resources/resources/cohortdefinition/sql/primaryEventsQuery.sql
@@ -1,8 +1,8 @@
 -- Begin Primary Events
-select row_number() over (PARTITION BY P.person_id order by P.start_date) as event_id, P.person_id, P.start_date, P.end_date, OP.observation_period_start_date as op_start_date, OP.observation_period_end_date as op_end_date, cast(P.visit_occurrence_id as int) as visit_occurrence_id
+select row_number() over (PARTITION BY P.person_id order by P.start_date) as event_id, P.person_id, P.start_date, P.end_date, OP.observation_period_start_date as op_start_date, OP.observation_period_end_date as op_end_date, cast(P.visit_occurrence_id as bigint) as visit_occurrence_id
 FROM
 (
-  select P.person_id, P.start_date, P.end_date, row_number() OVER (PARTITION BY person_id ORDER BY start_date @EventSort) ordinal, cast(P.visit_occurrence_id as int) as visit_occurrence_id
+  select P.person_id, P.start_date, P.end_date, row_number() OVER (PARTITION BY person_id ORDER BY start_date @EventSort) ordinal, cast(P.visit_occurrence_id as bigint) as visit_occurrence_id
   FROM 
   (
   @criteriaQueries
